+++ cypress/support/selectors.js
@@ -1,11 +1,5 @@
 export default {
   AddPickUpButton: '.w-70-ns > :nth-child(2) > .vtex-button__label',
-  PickUpPointName:
-    '.w-60-ns > .vtex-input > .vtex-input-prefix__group > .vtex-styleguide-9-x-input',
-  PickUpPointId:
-    '.pl5-ns > .vtex-input > .vtex-input-prefix__group > .vtex-styleguide-9-x-input',
-  PickUpAddress:
-    '.w-70-m > .vtex-input > .vtex-input-prefix__group > .vtex-styleguide-9-x-input',
   CheckBox:
     ':nth-child(3) > .w-40-m > .vtex-checkbox__line-container > .vtex-checkbox__container > .vtex-checkbox__input',
   WorkStartTime:
@@ -26,18 +20,6 @@ export default {
   VtexStoreName: 'span[class*=addressStoreName]',
   HoursContainer: 'div[class*=hoursContainer]',
   AddPickupPointButton: '.w-70-ns > :nth-child(2) > .vtex-button__label',
-  PickupPointName:
-    '.w-60-ns > .vtex-input > .vtex-input-prefix__group > .vtex-styleguide-9-x-input',
-  PickupPointId:
-    '.pl5-ns > .vtex-input > .vtex-input-prefix__group > .vtex-styleguide-9-x-input',
-  PickupPointAddress:
-    '.w-70-m > .vtex-input > .vtex-input-prefix__group > .vtex-styleguide-9-x-input',
-  SelectDay:
-    ':nth-child(3) > .w-40-m > .vtex-checkbox__line-container > .vtex-checkbox__container > .vtex-checkbox__input',
-  SelectHour:
-    ':nth-child(3) > .flex-ns > :nth-child(1) > .vtex-input > .vtex-input-prefix__group > .vtex-styleguide-9-x-input',
-  SelectMinutes:
-    ':nth-child(3) > .flex-ns > :nth-child(3) > .vtex-input > .vtex-input-prefix__group > .vtex-styleguide-9-x-input',
   SaveButton: '.flex.justify-end > .pb7 > .vtex-button > .vtex-button__label',
   StorePickUpPointList: '.overflow-auto > .list',
   PickupInStoresAddressLink: '.srp-address-title',