+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Increased hard retries and soft retries to verify inactive pickup points in cypress tests
+
+### Changed
+
 - Increased hard retries to verify pickup points in cypress tests
 
 ## [0.10.14] - 2022-10-07